% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroommlm.R
\docType{package}
\name{classroommlm}
\alias{classroommlm}
\title{classroommlm: A package for creating multilevel models from classroom data}
\section{Dependencies}{

The classroommlm package depends on the R2MlWin package, which is based on MLwiN. To use it:

\enumerate{
   \item Sign up for an academic account at \url{https://www.cmm.bristol.ac.uk/clients/reqform/}
   \item Download \code{mlnscript} for MacOS/linux by filling in form at \url{https://www.cmm.bristol.ac.uk/clients/softwaredownload/}
   \item Open the installer (.dmg, .rpm, etc) and save the files to a folder such as \code{/opt/mln}. (This is the path you will need to pass to the \code{\link{null_model}} and \code{\link{full_model}} functions.)
}
}

