% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{null_model}
\alias{null_model}
\title{Create the null model}
\usage{
null_model(pupil_data, mlwinpath)
}
\arguments{
\item{pupil_data}{a Data Frame with headings:
start_maths,student_id,class_id,N_in_class,Ability,Inattentiveness,hyper_impulsive,Deprivation,end_maths)}

\item{mlwinpath}{the path where mlnscript is installed (e.g. '/opt/mln/mlnscript')}
}
\value{
a Data Frame containing the summary data for the null model
}
\description{
This function creates a null model for the given pupil data.
}
