% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_mlm_with_features}
\alias{run_mlm_with_features}
\title{Create a full model excluding the Deprivation marker}
\usage{
run_mlm_with_features(pupil_data, mlwinpath, formula, features)
}
\arguments{
\item{pupil_data}{a Data Frame with headings:
start_maths,student_id,class_id,N_in_class,Ability,Inattentiveness,hyper_impulsive,Deprivation,end_maths)}

\item{mlwinpath}{the path where mlnscript is installed (e.g. '/opt/mln/mlnscript')}

\item{formula}{the formula to use for the mlwin model}

\item{features}{the features to use in the output}
}
\value{
a Data Frame containing the summary data for the full model
}
\description{
This function creates a full model for the given pupil data, not including
Deprivation
}
